+++ app/models/tarifs/price.rb
@@ -6,7 +6,7 @@
 #
 #  id                      :bigint           not null, primary key
 #  accountancy_account     :string
-#  invoice_types           :integer          default(0), not null
+#  associated_types        :integer          default(0), not null
 #  label_i18n              :jsonb
 #  minimum_usage_per_night :decimal(, )
 #  minimum_usage_total     :decimal(, )
@@ -15,7 +15,6 @@
 #  prefill_usage_method    :string
 #  price_per_unit          :decimal(, )
 #  tarif_group             :string
-#  tenant_visible          :boolean          default(TRUE)
 #  type                    :string
 #  unit_i18n               :jsonb
 #  valid_from              :datetime