+++ app/serializers/manage/tarif_serializer.rb
@@ -2,12 +2,11 @@
 
 module Manage
   class TarifSerializer < ApplicationSerializer
-    fields :label, :pin, :prefill_usage_method, :price_per_unit,
-           :tarif_group, :tenant_visible, :type, :unit, :ordinal,
+    fields :label, :pin, :prefill_usage_method, :price_per_unit, :tarif_group, :type, :unit, :ordinal,
            :label_i18n, :unit_i18n, :valid_from, :valid_until, :home_id
 
-    field :invoice_types do |tarif|
-      tarif.invoice_types.to_a
+    field :associated_types do |tarif|
+      tarif.associated_types.to_a
     end
 
     view :export do