+++ spec/factories/tarifs.rb
@@ -6,7 +6,7 @@
 #
 #  id                      :bigint           not null, primary key
 #  accountancy_account     :string
-#  invoice_types           :integer          default(0), not null
+#  associated_types        :integer          default(0), not null
 #  label_i18n              :jsonb
 #  minimum_usage_per_night :decimal(, )
 #  minimum_usage_total     :decimal(, )
@@ -15,7 +15,6 @@
 #  prefill_usage_method    :string
 #  price_per_unit          :decimal(, )
 #  tarif_group             :string
-#  tenant_visible          :boolean          default(TRUE)
 #  type                    :string
 #  unit_i18n               :jsonb
 #  valid_from              :datetime
@@ -43,7 +42,7 @@ FactoryBot.define do
     unit { 'Übernachtung (unter 16 Jahren)' }
     price_per_unit { 15.0 }
     organisation
-    invoice_types { Tarif.invoice_types.keys }
+    associated_types { Tarif.associated_types.keys }
     prefill_usage_method { nil }
   end
 end